Reimplement EventStream in a pull-based manner
The pull-based implementation is a better fit for the patterns in
Essential Scala. We still run into GADTs but its worthwhile introducing
them as I expect will run into them on their own, and need to know how
to handle them.

The pull-based implementation has better semantics than the push-based
implementation. It respects substitution and allows to to enqueue events
without the messy hot and cold streams that Rx implementations have.
This is achieved by, as always, separating the structure of the
computation from the process that gives it meaning. The former is the
EventStream type. This is then compiled into an intermediate
representation, which can then be run.

It currently uses a busy poll; next step is to replace
this with sources pushing to a queue when events are ready. This will
reduce CPU consumption.

Consider changing the IR name. It's not particularly descriptive.